+++ resources/visu_vtk/src/lib/vtk.d.ts
@@ -1,1 +1,5 @@
 declare const vtk: any;
+
+interface Window {
+  vtk: any;
+}